Fix type widening

+++ casper/src/main/scala/coop/rchain/casper/util/comm/DeployRuntime.scala
@@ -71,7 +71,7 @@ object DeployRuntime {
   def deployFileProgram[F[_]: Monad: Sync: DeployService](
       phloLimit: Long,
       phloPrice: Long,
-      validAfterBlock: Int,
+      validAfterBlock: Long,
       maybePrivateKey: Option[PrivateKey],
       file: String
   ): F[Unit] =
